More PowerShell bug fixes

+++ Src/Scripts/setup-agent-host.ps1
@@ -238,10 +238,14 @@ process {
         # Clear build agent working directory
         [string]$workingDirectory = [System.IO.Path]::Combine($workDirectory, "B")
 
-        # Work around PowerShell bugs: https://github.com/powershell/powershell/issues/621
-        gci $workingDirectory -Recurse -Attributes ReparsePoint | % { $_.Delete() }
+        # If the path doesn't exist Get-ChildItem will happily pick the working directory instead which could delete C:\Windows\ ...
+        # https://github.com/PowerShell/PowerShell/issues/5699
+        if (Test-Path $workingDirectory) {            
+            # Work around PowerShell bugs: https://github.com/powershell/powershell/issues/621
+            Get-ChildItem -LiteralPath $workingDirectory -Recurse -Attributes ReparsePoint | % { $_.Delete() }
         
-        Remove-Item -Path $workingDirectory -Force -Recurse -Verbose -ErrorAction SilentlyContinue
+            Remove-Item -Path $workingDirectory -Force -Recurse -Verbose -ErrorAction SilentlyContinue 
+        }
         
         & $PSScriptRoot\iis-cleanup.ps1
